requirements.txt
streamlit==1.17.0
fastapi==0.89.1
lightgbm==3.3.2
matplotlib==3.5.1
numpy==1.23.4
pandas==1.3.5
pytest==7.2.1
PyYAML==6.0
ray==2.2.0
requests==2.27.1
scikit_learn==1.2.1
seaborn==0.11.2
shap==0.41.0
tqdm==4.64.1
uvicorn==0.20.0
virtualenv==20.17.1
joblib




